+++ src/main/java/c4/champions/common/affix/AffixEvents.java
@@ -48,13 +48,9 @@ public class AffixEvents {
 
             if (chp != null) {
 
-                for (String aff : chp.getAffixes()) {
-                    IAffix affix = EnumAffix.getAffix(aff);
-
-                    if (affix != null) {
-                        affix.onSpawn(living, chp);
-                    }
-                }
+                chp.getAffixes().stream().forEach(ordinal ->
+                    EnumAffix.getAffix(ordinal).onSpawn(living, chp)
+                );
                 Rank rank = chp.getRank();
 
                 if (!living.world.isRemote && ChampionHelper.isElite(rank)) {
@@ -79,14 +75,9 @@ public class AffixEvents {
             IChampionship chp = CapabilityChampionship.getChampionship(living);
 
             if (chp != null) {
-
-                for (String aff : chp.getAffixes()) {
-                    IAffix affix = EnumAffix.getAffix(aff);
-
-                    if (affix != null) {
-                        affix.onUpdate(living, chp);
-                    }
-                }
+                chp.getAffixes().stream().forEach(ordinal ->
+                    EnumAffix.getAffix(ordinal).onUpdate(living, chp)
+                );
                 Rank rank = chp.getRank();
 
                 if (ChampionHelper.isElite(rank)) {
@@ -116,14 +107,9 @@ public class AffixEvents {
             IChampionship chp = CapabilityChampionship.getChampionship(living);
 
             if (chp != null) {
-
-                for (String aff : chp.getAffixes()) {
-                    IAffix affix = EnumAffix.getAffix(aff);
-
-                    if (affix != null) {
-                        affix.onAttacked(living, chp, evt.getSource(), evt.getAmount(), evt);
-                    }
-                }
+                chp.getAffixes().stream().forEach(ordinal ->
+                    EnumAffix.getAffix(ordinal).onAttacked(living, chp, evt.getSource(), evt.getAmount(), evt)
+                );
             }
         }
     }
@@ -139,13 +125,9 @@ public class AffixEvents {
 
                 if (chp != null) {
 
-                    for (String aff : chp.getAffixes()) {
-                        IAffix affix = EnumAffix.getAffix(aff);
-
-                        if (affix != null) {
-                            affix.onAttack(living, chp, evt.getEntityLiving(), evt.getSource(), evt.getAmount(), evt);
-                        }
-                    }
+                    chp.getAffixes().stream().forEach(ordinal ->
+                        EnumAffix.getAffix(ordinal).onAttack(living, chp, evt.getEntityLiving(), evt.getSource(), evt.getAmount(), evt)
+                    );
                 }
             }
         }
@@ -156,20 +138,19 @@ public class AffixEvents {
 
         if (ChampionHelper.isValidChampion(evt.getEntityLiving())) {
             float amount = evt.getAmount();
-            float newAmount = amount;
+            float newAmount;
 
             EntityLiving living = (EntityLiving)evt.getEntityLiving();
             IChampionship chp = CapabilityChampionship.getChampionship(living);
 
             if (chp != null) {
 
-                for (String aff : chp.getAffixes()) {
-                    IAffix affix = EnumAffix.getAffix(aff);
-
-                    if (affix != null) {
-                        newAmount = affix.onHurt(living, chp, evt.getSource(), amount, newAmount);
-                    }
-                }
+                newAmount = chp.getAffixes().stream()
+                    .mapToObj(ordinal -> EnumAffix.getAffix(ordinal).onHurt(living, chp, evt.getSource(), amount, amount))
+                    .filter(f -> f != amount)
+                    .findAny()
+                    .orElse(amount)
+                ;
                 evt.setAmount(newAmount);
             }
         }
@@ -180,20 +161,18 @@ public class AffixEvents {
 
         if (ChampionHelper.isValidChampion(evt.getEntityLiving())) {
             float amount = evt.getAmount();
-            float newAmount = amount;
+            float newAmount;
 
             EntityLiving living = (EntityLiving)evt.getEntityLiving();
             IChampionship chp = CapabilityChampionship.getChampionship(living);
 
             if (chp != null) {
-
-                for (String aff : chp.getAffixes()) {
-                    IAffix affix = EnumAffix.getAffix(aff);
-
-                    if (affix != null) {
-                        newAmount = affix.onDamaged(living, chp, evt.getSource(), amount, newAmount);
-                    }
-                }
+                newAmount = chp.getAffixes().stream()
+                    .mapToObj(ordinal -> EnumAffix.getAffix(ordinal).onDamaged(living, chp, evt.getSource(), amount, amount))
+                    .filter(f -> f != amount)
+                    .findAny()
+                    .orElse(amount)
+                ;
                 evt.setAmount(newAmount);
             }
         }
@@ -207,14 +186,9 @@ public class AffixEvents {
             IChampionship chp = CapabilityChampionship.getChampionship(living);
 
             if (chp != null) {
-
-                for (String aff : chp.getAffixes()) {
-                    IAffix affix = EnumAffix.getAffix(aff);
-
-                    if (affix != null) {
-                        affix.onKnockback(living, chp, evt);
-                    }
-                }
+                chp.getAffixes().stream().forEach(ordinal ->
+                    EnumAffix.getAffix(ordinal).onKnockback(living, chp, evt)
+                );
             }
         }
     }
@@ -228,13 +202,9 @@ public class AffixEvents {
 
             if (chp != null) {
 
-                for (String aff : chp.getAffixes()) {
-                    IAffix affix = EnumAffix.getAffix(aff);
-
-                    if (affix != null) {
-                        affix.onDeath(living, chp, evt.getSource(), evt);
-                    }
-                }
+                chp.getAffixes().stream().forEach(ordinal ->
+                    EnumAffix.getAffix(ordinal).onDeath(living, chp, evt.getSource(), evt)
+                );
             }
         }
     }
@@ -244,20 +214,18 @@ public class AffixEvents {
 
         if (ChampionHelper.isValidChampion(evt.getEntityLiving())) {
             float amount = evt.getAmount();
-            float newAmount = amount;
+            float newAmount;
 
             EntityLiving living = (EntityLiving)evt.getEntityLiving();
             IChampionship chp = CapabilityChampionship.getChampionship(living);
 
             if (chp != null) {
-
-                for (String aff : chp.getAffixes()) {
-                    IAffix affix = EnumAffix.getAffix(aff);
-
-                    if (affix != null) {
-                        newAmount = affix.onHealed(living, chp,  amount, newAmount);
-                    }
-                }
+                newAmount = chp.getAffixes().stream()
+                    .mapToObj(ordinal -> EnumAffix.getAffix(ordinal).onHealed(living, chp,  amount, amount))
+                    .filter(f -> f != amount)
+                    .findAny()
+                    .orElse(amount)
+                ;
                 evt.setAmount(newAmount);
             }
         }